version 21.4.49.0

+++ OpenDentBusiness/TableTypes/CovCat.cs
@@ -14,7 +14,7 @@ namespace OpenDentBusiness{
 		///<summary>Default percent for this category. -1 to skip this category and not apply a percentage.</summary>
 		public int DefaultPercent;
 		///<summary>The order in which the categories are displayed.  Includes hidden categories. 0-based.</summary>
-		public byte CovOrder;
+		public int CovOrder;
 		///<summary>If true, this category will be hidden.</summary>
 		public bool IsHidden;
 		///<summary>Enum:EbenefitCategory  The X12 benefit categories.  Each CovCat can link to one X12 category.  Default is 0 (unlinked).</summary>
